refactor: move context outside of the hooks folder
